pyproject.toml
[project]
name = "SBI 電子交付書面 download"
version = "v0.1.0"
description = ""
authors = [
    {name = "Yufei Kang", email = "kou.uhi.x@gmail.com"},
]
dependencies = [
    "pytest-playwright>=0.3.0",
    "python-dotenv>=0.21.1",
    "greenlet>=1.1.2",
]
requires-python = ">=3.9"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.6.0",
]


[tool.pdm.scripts]
sbi = "python sbi.py"
post_install = "pdm run playwright install chromium"
